+++ docs/CHANGELOG.md
@@ -7,3 +7,4 @@
 + Add ST_ConcaveHull function
 + Improve ST_Graph function to keep a list of columns for the edges
 + Add ST_LineSubstring and ST_LineInterpolatePoint functions
++ Add ST_MaximumInscribedCircle function